fix build

+++ dbms/src/Flash/Planner/Plans/PhysicalJoin.cpp
@@ -122,7 +122,7 @@ PhysicalPlanNodePtr PhysicalJoin::build(
 
     const Settings & settings = context.getSettingsRef();
     // Currently, the pipeline model does not support disk-based joins, so when force_enable_pipeline is true, the disk-based join will be disabled.
-    size_t max_bytes_before_external_join = settings.force_enable_pipeline ? 0 : settings.max_bytes_before_external_join;
+    size_t max_bytes_before_external_join = settings.force_enable_pipeline ? 0 : static_cast<size_t>(settings.max_bytes_before_external_join);
     SpillConfig build_spill_config(context.getTemporaryPath(), fmt::format("{}_hash_join_0_build", log->identifier()), settings.max_cached_data_bytes_in_spiller, settings.max_spilled_rows_per_file, settings.max_spilled_bytes_per_file, context.getFileProvider());
     SpillConfig probe_spill_config(context.getTemporaryPath(), fmt::format("{}_hash_join_0_probe", log->identifier()), settings.max_cached_data_bytes_in_spiller, settings.max_spilled_rows_per_file, settings.max_spilled_bytes_per_file, context.getFileProvider());
     size_t max_block_size = settings.max_block_size;
